chore(tests): migrate jsii-spec tests to jest
In an effort to try and stop using `nodeunit` everywhere, migrating all
existing tests in the `jsii-spec` module to `jest`.
