+++ packages/jsii-spec/lib/assembly.ts
