+++ packages/jsii-spec/lib/index.ts
@@ -1,3 +1,3 @@
+export * from './assembly';
 export * from './name-tree';
-export * from './spec';
 export * from './validate-assembly';
