+++ packages/jsii-spec/package.json
@@ -7,18 +7,40 @@
   "scripts": {
     "build": "tsc --build && bash generate-json-schema.sh",
     "watch": "tsc --build -w",
-    "test": "nodeunit test/test.*.js",
+    "test": "jest",
     "package": "package-js"
   },
   "devDependencies": {
-    "@types/nodeunit": "^0.0.30",
+    "@types/jest": "^24.0.18",
+    "jest": "^24.9.0",
     "jsii-build-tools": "file:../jsii-build-tools",
-    "nodeunit": "^0.11.3",
     "typescript": "^3.5.3",
     "typescript-json-schema": "^0.40.0"
   },
   "dependencies": {
     "jsonschema": "^1.2.4"
+  },
+  "jest": {
+    "collectCoverage": true,
+    "collectCoverageFrom": [
+      "**/bin/**/*.js",
+      "**/lib/**/*.js"
+    ],
+    "coverageReporters": [
+      "lcov",
+      "text"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "branches": 35,
+        "statements": 60
+      }
+    },
+    "errorOnDeprecated": true,
+    "testEnvironment": "node",
+    "testMatch": [
+      "**/?(*.)+(spec|test).js"
+    ]
   },
   "author": {
     "name": "Amazon Web Services",
