+++ packages/jsii-spec/test/validate-assembly.test.ts
@@ -1,0 +1,6 @@
+import { validateAssembly } from '../lib/validate-assembly';
+
+describe('validateAssembly', () => {
+    test('rejects invalid assembly', () =>
+        expect(() => validateAssembly({})).toThrow(/Invalid assembly:/));
+});
